+++ lib/helpers/tag_emoji.dart
@@ -28,6 +28,7 @@ const kTagEmoji = <String, String>{
   'amenity=dentist': '🦷',
   'amenity=bus_station': '🚏',
   'amenity=police': '🚓',
+  'amenity=bureau_de_change': '💱',
 
   'shop=convenience': '🛒',
   'shop=supermarket': '🛒',
@@ -45,12 +46,22 @@ const kTagEmoji = <String, String>{
   'shop=shoes': '👞',
   'shop=car_repair': '🚗',
   'shop=car_parts': '🚗',
-  'shop=bakery': '🍞',
+
+  'shop=bakery': '🥨',
+  'shop=pastry': '🍰',
   'shop=butcher': '🥩',
   'shop=furniture': '🛋️',
   'shop=mobile_phone': '📱',
+  'shop=tobacco': '🚬',
+  'shop=jewelry': '💎',
+  'shop=fashion_accessories': '📿',
+  'shop=cosmetics': '💄',
+  'shop=sports': '⚽',
+  'shop=optician': '👓',
   'shop=mall': '🛍️',
   'shop=ice_cream': '🍨',
+
+  'craft=electronics_repair': '🔌',
 
   'tourism=information': 'ℹ️',
   'tourism=hotel': '🏨',