+++ CHANGELOG.md
@@ -8,18 +8,21 @@ _Not yet released_
 * Fixed OAuth error with a `FormatException`.
 * First upload after app start redirected to the login page.
 * Added address and floor filters.
+* Added a raw tag editor.
 * Pending changes can be deleted from the list by swiping left.
+* Now can delete polygon amenities (using `was:` prefix).
 * Moved "disused" and "missing" buttons below the fields.
 * Removed the "missing" button for freshly created amenities.
 * Better message for when there are no POI around.
 * Changed app name to "Every Door" from "every\_door".
 * Radio buttons stay in place on tap when they fit the screen.
-* Removed `shop`, `building_area`, and `opening_hours/covid19` fields.
+* Removed `building_area` and `opening_hours/covid19` fields.
 * Types now can be found by tag values.
 * It was possible to have an opening\_hours fragment with no weekdays.
+* German translation (thanks @mfbehrens99).
 * With GPS disabled, app now restores the last location.
-* Fixed empty floor list when adding an address to a POI with floor tags.
 * Fixed a map error when GPS is off.
+* Fixed empty floor list when adding an address to a POI with floor tags.
 * Fixed a database error when restoring multipolygon relations.
 
 ## 0.1.0